+++ packages/app/hivecommand-backend/src/schema/schematics/index.ts
@@ -7,6 +7,7 @@ import { subject } from "@casl/ability";
 
 import { LexoRank } from 'lexorank';
 import { InvokeCommand, LambdaClient } from '@aws-sdk/client-lambda'
+import moment from "moment";
 
 export default (prisma: PrismaClient) => {
 
@@ -14,7 +15,11 @@ export default (prisma: PrismaClient) => {
 
 	const resolvers = mergeResolvers([
 		{
-
+			CommandSchematicVersion: {
+				createdBy: (root: any) => {
+					return {id: root.createdBy}
+				}
+			},
 			Query: {
 				commandSchematics: async (root: any, args: any, context: any) => {
 
@@ -24,7 +29,8 @@ export default (prisma: PrismaClient) => {
 					const programs = await prisma.electricalSchematic.findMany({
 						where: { ...filter, organisation: context.jwt.organisation },
 						include: {
-							pages: true
+							pages: true,
+							versions: true
 						}
 					});
 
@@ -80,7 +86,36 @@ export default (prisma: PrismaClient) => {
 					const res = await prisma.electricalSchematic.delete({ where: { id: args.id } })
 					return res != null
 				},
+				createCommandSchematicVersion: async (root: any, args: any, context: any) => {
 
+					const schematic = await prisma.electricalSchematic.findFirst({
+						where: {
+							id: args.id, 
+							organisation: context?.jwt?.organisation
+						},
+						include: {
+							pages: true,
+							versions: true
+						}
+					})
+
+					const {versions} = schematic || {};
+
+					const nextVersion = (versions?.sort((a, b) => b.rank - a.rank)?.[0]?.rank || 0) + 1;
+
+					const version = await prisma.electricalSchematicVersion.create({
+						data: {
+							id: nanoid(),
+							data: schematic || {},
+							rank: nextVersion,
+							commit: args.input.commit,
+							createdBy: context?.jwt?.id,
+							schematic: {connect: {id: args.id}}
+						}
+					})
+
+					return nextVersion + 1;
+				},
 				exportCommandSchematic: async (root: any, args: { id: string }, context: any) => {
 					const currentProgram = await prisma.electricalSchematic.findFirst({
 						where: {
@@ -88,14 +123,17 @@ export default (prisma: PrismaClient) => {
 							organisation: context?.jwt?.organisation
 						},
 						include: {
-							pages: true
+							pages: true,
+							versions: true
 						}
 					})
+
+					const version = currentProgram?.versions?.sort((a, b) => b.rank - a.rank)?.[0];
 
 					const invokeCommand = new InvokeCommand({
 						FunctionName: process.env.EXPORT_LAMBDA || '',
 						Payload: JSON.stringify({
-							program: currentProgram
+							program: { ...currentProgram, version: (version?.rank || 0) + 1, versionDate: moment(version?.createdAt).format('DD/MM/YY') }
 						})
 					})
 
@@ -135,8 +173,8 @@ export default (prisma: PrismaClient) => {
 						data: {
 							id: nanoid(),
 							name: args.input.name,
-							nodes: args.input.nodes,
-							edges: args.input.edges,
+							nodes: args.input.nodes || [],
+							edges: args.input.edges || [],
 							rank: newRank.toString(),
 							schematic: {
 								connect: { id: args.schematic }
@@ -232,6 +270,7 @@ export default (prisma: PrismaClient) => {
 		updateCommandSchematic(id: ID!, input: CommandSchematicInput!): CommandSchematic!
 		deleteCommandSchematic(id: ID!): Boolean!
 
+		createCommandSchematicVersion(id: ID!, input: CommandSchematicVersionInput): String
 		exportCommandSchematic(id: ID!): String
 
         createCommandSchematicPage(schematic: ID, input: CommandSchematicPageInput): CommandSchematicPage!
@@ -257,12 +296,33 @@ export default (prisma: PrismaClient) => {
 	type CommandSchematic {
 		id: ID! 
 		name: String
+		versions: [CommandSchematicVersion]
 
         pages: [CommandSchematicPage]
 
 		createdAt: DateTime 
 
 		organisation: HiveOrganisation
+	}
+
+	input CommandSchematicVersionInput {
+		commit: String
+	}
+
+	type CommandSchematicVersion {
+		id: ID
+	  
+		rank: Int
+	  
+		data: JSON
+	  
+		commit: String
+	  
+		createdAt: DateTime
+	  
+		createdBy: HiveUser
+	  
+		schematic: CommandSchematic 
 	}
 
     input CommandSchematicPageInput{