+++ packages/core-ui/command-electrical-editor/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.271](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.270...v1.4.15-alpha.271) (2023-10-09)
+
+**Note:** Version bump only for package @hive-command/electrical-editor
+
+
+
+
+
+## [1.4.15-alpha.270](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.269...v1.4.15-alpha.270) (2023-10-09)
+
+**Note:** Version bump only for package @hive-command/electrical-editor
+
+
+
+
+
+## [1.4.15-alpha.269](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.268...v1.4.15-alpha.269) (2023-10-09)
+
+**Note:** Version bump only for package @hive-command/electrical-editor
+
+
+
+
+
+## [1.4.15-alpha.268](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.267...v1.4.15-alpha.268) (2023-10-09)
+
+**Note:** Version bump only for package @hive-command/electrical-editor
+
+
+
+
+
 ## [1.4.15-alpha.267](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.266...v1.4.15-alpha.267) (2023-09-12)
 
 **Note:** Version bump only for package @hive-command/electrical-editor